// src/app/cars/models/car.ts
export interface Car {
    id: number;
    model: string;
    type: string;
    year: string;
    color: string;
    cost: number;
    isFullyDamaged: boolean;
    clientFirstName: string;
    clientSurname: string;
    power: number;
    plate: string;
    deliveryDate: string;
    deadline: string;
}